% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_apneas.R
\name{compute_odi}
\alias{compute_odi}
\title{Compute Oxygen Desaturation Index (ODI)}
\usage{
compute_odi(psg, spo2_channel, tst_hours, threshold = 3)
}
\arguments{
\item{psg}{An \code{mrpheus_psg} object from \code{\link[=prepare_psg]{prepare_psg()}}.}

\item{spo2_channel}{Character. SpO2 channel label.}

\item{tst_hours}{Numeric. Total sleep time in hours.}

\item{threshold}{Numeric. Desaturation threshold (percentage points).
Default \code{3}.}
}
\value{
Numeric scalar. ODI (desaturations per hour).
}
\description{
Calculates the ODI (number of >= 3\% SpO2 desaturations per hour of sleep)
from the SpO2 channel.
}
