% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpheus-package.R
\docType{package}
\name{mrpheus-package}
\alias{mrpheus}
\alias{mrpheus-package}
\title{mrpheus: Polysomnography Signal Analysis for Sleep Research}
\description{
Raw polysomnography (PSG) signal analysis for sleep and circadian research. Provides EDF/EDF+ ingestion, artefact detection, spectral analysis, sleep event detection (spindles, slow oscillations), automatic AASM sleep staging via a pre-trained LightGBM model (ported from YASA; Vallat & Walker, 2021), and respiratory and cardiac metrics. Exports staged hypnograms directly to hypnor and PSG-derived metrics to syncR. Part of the Circadia Lab R ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/circadia-bio/mrpheus}
  \item Report bugs at \url{https://github.com/circadia-bio/mrpheus/issues}
}

}
\author{
\strong{Maintainer}: Lucas França \email{lucas.franca@northumbria.ac.uk}

Authors:
\itemize{
  \item Lucas França \email{lucas.franca@northumbria.ac.uk}
  \item Mario Leocadio-Miguel (\href{https://orcid.org/0000-0002-7248-3529}{ORCID})
}

}
\keyword{internal}
